pyproject.toml
[project]
name = "outskill-ai-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "openai-agents>=0.6.5",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.17",
]
